# openapi.yaml
openapi: 3.1.0
info:
  title: Grainlify API
  description: |
    The Grainlify backend API. Public endpoints are accessible without authentication.
    Authenticated endpoints require a Bearer JWT token obtained via `/auth/verify`.
    Admin endpoints additionally require the `admin` role.
  version: 1.0.0
  contact:
    name: Grainlify OSS
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.grainlify.0xo.in
    description: Production
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: The requested resource was not found
      required: [error]
...
    KYCStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [none, pending, approved, rejected]
        updated_at:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /auth/kyc/status:
    get:
      summary: Get KYC verification status
      operationId: kycStatus
      tags: [Auth, KYC]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: KYC status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /webhooks/didit:
    get:
      summary: Didit KYC callback redirect
      operationId: receiveDiditWebhookGet
      tags: [Webhooks]
      responses:
        '302':
          description: Redirect after KYC completion
    post:
      summary: Receive Didit KYC webhook events
      operationId: receiveDiditWebhookPost
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Didit webhook payload
      responses:
        '200':
          description: Webhook received
...